Fix crash in some cases when fusions had int4 inputs/outputs.
Before, if a fusion had an int8 input and int4 output, or vice versa, and the int4 tensor was not a ROOT or parameter of the entry computation, a crash could occur. The reason is the FusionCanShareBufferHint would incorrectly use ByteSizeOfPrimitiveType instead of BitWidth to check if two buffers were the same size.

I found this through an autogenerated HLO with XLA smith, to test an unrelated change that adds int4 support to most ops. Unfortunately the HLO, which is added as a test case, is pretty weird. It has a dot with no contracting or non-contracting dimensions, just consisting of batch dimensions. The issue also only occurred with algebraic simplifier disabled.

I couldn't find any simpler HLO that had a fusion with an int8 input and int4 output such that the int4 tensor was not a ROOT instruction. There must always be a convert back from int4 to int8 (as the int4 tensor cannot be a ROOT instruction), and any HLO I created would always have the fusion include the convert back to int8.

PiperOrigin-RevId: 687132941